backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic>=2.5.3
pydantic-settings==2.1.0
pyyaml==6.0.1
python-multipart==0.0.6
aiofiles==23.2.1
requests==2.31.0
urllib3==2.1.0
